tracker/pt: all callers are double, so float -> double

+++ tracker-pt/camera.h
@@ -41,7 +41,7 @@ public:
         void set_res(int x_res, int y_res);
 
         // gets a frame from the camera, dt: time since last call in seconds
-        bool get_frame(float dt, cv::Mat* frame);
+        bool get_frame(double dt, cv::Mat* frame);
 
         // WARNING: returned references are valid as long as object
         bool get_info(CamInfo &ret);
@@ -57,8 +57,8 @@ protected:
         virtual void _set_fps() = 0;
         virtual void _set_res() = 0;
 private:
-        float dt_valid;
-        float dt_mean;
+        double dt_valid;
+        double dt_mean;
 protected:
         CamInfo cam_info;
         CamInfo cam_desired;
